requirements_multi_agent.txt
# Multi-Agent Trading Bot Requirements
# Enhanced version with specialized agents and optimizations

# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=6.0
python-dotenv>=0.19.0

# Async and networking
aiohttp>=3.8.0
asyncio-throttle>=1.0.2
urllib3>=1.26.0
requests>=2.28.0

# Binance API
python-binance>=1.0.15

# Machine Learning & RL
tensorflow-cpu==2.19.0
stable-baselines3>=2.6.0
scikit-learn>=1.1.0
gymnasium>=1.1.1

# Technical Analysis
TA-Lib>=0.4.25  # Requires manual C library installation
pandas-ta>=0.3.14b  # Fallback for TA-Lib

# Sentiment Analysis
transformers>=4.21.0
torch>=2.7.1  # CPU-only version
onnxruntime>=1.22.0
huggingface-hub>=0.10.0

# Social Media APIs
praw>=7.6.0  # Reddit
tweepy>=4.14.0  # Twitter (optional)

# Telegram Bot
python-telegram-bot>=20.0

# Data Processing & Caching
redis>=4.3.0  # Optional: for distributed caching
lz4>=4.0.0  # Compression
msgpack>=1.0.4  # Serialization

# Monitoring & Logging
psutil>=5.9.0  # System monitoring
prometheus-client>=0.14.0  # Optional: metrics collection

# Development & Testing
pytest>=7.1.0
pytest-asyncio>=0.19.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.971

# Optional: Database support
sqlalchemy>=1.4.0  # Optional: for trade history
sqlite3  # Built-in, for local DB

# Optional: Web interface
flask>=2.2.0  # Optional: for web dashboard
flask-cors>=3.0.0

# Visualization (optional)
matplotlib>=3.5.0
plotly>=5.10.0

# Performance profiling (development)
py-spy>=0.3.12
memory-profiler>=0.60.0

# Security
cryptography>=37.0.0

# Utilities
click>=8.1.0  # CLI interface
rich>=12.5.0  # Pretty terminal output
tqdm>=4.64.0  # Progress bars